+++ packages/npm/@amazeelabs/publisher/rollup.config.js
@@ -1,3 +1,4 @@
+import alias from '@rollup/plugin-alias';
 import commonjs from '@rollup/plugin-commonjs';
 import json from '@rollup/plugin-json';
 import resolve from '@rollup/plugin-node-resolve';
@@ -28,5 +29,20 @@ export default [
       format: 'iife',
     },
     plugins: [litcss(), esbuild(), resolve()],
+  },
+  {
+    input: `src/elements.ts`,
+    output: {
+      file: `dist/elements.js`,
+      format: 'iife',
+    },
+    plugins: [
+      alias({
+        entries: [{ find: 'lodash', replacement: 'lodash-es' }],
+      }),
+      litcss(),
+      esbuild(),
+      resolve(),
+    ],
   },
 ];