+++ packages/npm/@amazeelabs/react-framework-bridge/src/gatsby.tsx
@@ -8,6 +8,8 @@ import { Form, FormBuilderProps } from './types';
 import {
   buildHtmlBuilder,
   buildUrlBuilder,
+  FormikChanges,
+  FormikInitialValues,
   isInternalTarget,
   isRelative,
 } from './utils';
@@ -115,10 +117,12 @@ export const buildImage = (props: ImageProps): Image => {
 
 export const buildHtml = buildHtmlBuilder(buildLink);
 
-export function buildForm<Values>(
-  formikProps: FormBuilderProps<Values>,
-): Form<Values> {
-  return function GatsbyFormBuilder(formProps) {
+export function buildForm<Values>({
+  onChange,
+  useInitialValues,
+  ...formikProps
+}: FormBuilderProps<Values>): Form<Values> {
+  return function GatsbyFormBuilder({ children, ...formProps }) {
     return (
       <Formik
         onSubmit={(values) => {
@@ -126,7 +130,13 @@ export function buildForm<Values>(
         }}
         {...formikProps}
       >
-        <FormComponent {...formProps} />
+        <FormComponent {...formProps}>
+          {onChange ? <FormikChanges onChange={onChange} /> : null}
+          {useInitialValues ? (
+            <FormikInitialValues useInitialValues={useInitialValues} />
+          ) : null}
+          {children}
+        </FormComponent>
       </Formik>
     );
   };