+++ packages/composer/drupal/gatsby_build_monitor/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.27](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.4.26...@-drupal/gatsby_build_monitor@1.4.27) (2022-03-18)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
+## [1.4.26](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.4.25...@-drupal/gatsby_build_monitor@1.4.26) (2022-03-18)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
+## [1.4.25](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.4.24...@-drupal/gatsby_build_monitor@1.4.25) (2022-03-10)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
 ## [1.4.24](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.4.23...@-drupal/gatsby_build_monitor@1.4.24) (2022-03-03)
 
 **Note:** Version bump only for package @-drupal/gatsby_build_monitor