+++ packages/npm/@amazeelabs/react-di/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.57](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.56...@amazeelabs/react-di@1.0.57) (2022-03-18)
+
+**Note:** Version bump only for package @amazeelabs/react-di
+
+
+
+
+
+## [1.0.56](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.55...@amazeelabs/react-di@1.0.56) (2022-03-18)
+
+**Note:** Version bump only for package @amazeelabs/react-di
+
+
+
+
+
+## [1.0.55](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.54...@amazeelabs/react-di@1.0.55) (2022-03-10)
+
+**Note:** Version bump only for package @amazeelabs/react-di
+
+
+
+
+
 ## [1.0.54](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-di@1.0.53...@amazeelabs/react-di@1.0.54) (2022-03-03)
 
 **Note:** Version bump only for package @amazeelabs/react-di