+++ packages/npm/@amazeelabs/silverback-iframe/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.33](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-iframe@1.1.32...@amazeelabs/silverback-iframe@1.1.33) (2022-03-18)
+
+**Note:** Version bump only for package @amazeelabs/silverback-iframe
+
+
+
+
+
+## [1.1.32](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-iframe@1.1.31...@amazeelabs/silverback-iframe@1.1.32) (2022-03-18)
+
+**Note:** Version bump only for package @amazeelabs/silverback-iframe
+
+
+
+
+
+## [1.1.31](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-iframe@1.1.30...@amazeelabs/silverback-iframe@1.1.31) (2022-03-10)
+
+**Note:** Version bump only for package @amazeelabs/silverback-iframe
+
+
+
+
+
 ## [1.1.30](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-iframe@1.1.29...@amazeelabs/silverback-iframe@1.1.30) (2022-03-03)
 
 **Note:** Version bump only for package @amazeelabs/silverback-iframe