+++ packages/npm/@amazeelabs/silverback-playwright/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.23](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.6.22...@amazeelabs/silverback-playwright@1.6.23) (2022-03-18)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.6.22](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.6.21...@amazeelabs/silverback-playwright@1.6.22) (2022-03-18)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.6.21](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.6.20...@amazeelabs/silverback-playwright@1.6.21) (2022-03-10)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
 ## [1.6.20](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.6.19...@amazeelabs/silverback-playwright@1.6.20) (2022-03-03)
 
 **Note:** Version bump only for package @amazeelabs/silverback-playwright