+++ packages/tests/silverback-gatsby/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.38](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gatsby-tests@1.1.37...@-amazeelabs/silverback-gatsby-tests@1.1.38) (2022-03-18)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gatsby-tests
+
+
+
+
+
+## [1.1.37](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gatsby-tests@1.1.36...@-amazeelabs/silverback-gatsby-tests@1.1.37) (2022-03-18)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gatsby-tests
+
+
+
+
+
+## [1.1.36](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gatsby-tests@1.1.35...@-amazeelabs/silverback-gatsby-tests@1.1.36) (2022-03-15)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gatsby-tests
+
+
+
+
+
+## [1.1.35](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gatsby-tests@1.1.34...@-amazeelabs/silverback-gatsby-tests@1.1.35) (2022-03-10)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-gatsby-tests
+
+
+
+
+
 ## [1.1.34](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-gatsby-tests@1.1.33...@-amazeelabs/silverback-gatsby-tests@1.1.34) (2022-03-08)
 
 **Note:** Version bump only for package @-amazeelabs/silverback-gatsby-tests